cleanup: move validatePhoneNumber from MembershipController to MemberManager
rationale: WsController must not depend on other controllers (needed to use spring MVC instead of jersey)
